pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "flask-login",
    "flask-wtf>=1.2.1",
    "scikit-learn>=1.5.2",
    "numpy>=2.1.2",
]
